fixed wrong data type of INTEGER fill_value in gather and exchange routines (thanks to J. Riehme). (fill_value feature currently unused for INTEGER).

+++ src/parallel_infrastructure/mo_communication.f90
@@ -1215,7 +1215,7 @@ CONTAINS
     INTEGER, INTENT(IN) :: in_array(:,:)
     ! dimension (global length); only required on root
     INTEGER, INTENT(INOUT) :: out_array(:)
-    REAL(dp), INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
+    INTEGER, INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
     ! be replaced with this value
     ! if not provided all valid
     ! points will be packed to the
@@ -1337,7 +1337,7 @@ CONTAINS
     INTEGER, INTENT(IN) :: in_array(:,:,:)
     ! dimension (global length, nlev); only required on root
     INTEGER, INTENT(INOUT) :: out_array(:,:)
-    REAL(dp), INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
+    INTEGER, INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
     ! be replaced with this value
     ! if not provided all valid
     ! points will be packed to the
@@ -1462,7 +1462,7 @@ CONTAINS
     INTEGER, INTENT(IN) :: in_array(:,:)
     ! dimension (global length); only required on root
     INTEGER, INTENT(INOUT) :: out_array(:)
-    REAL(dp), INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
+    INTEGER, INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
     ! be replaced with this value
     ! if not provided all valid
     ! points will be packed to the
@@ -1705,7 +1705,7 @@ CONTAINS
     &                                 gather_pattern, collector_buffer_i)
     ! dimension (:, length), prepared according to gather pattern
     INTEGER, INTENT(IN) :: send_buffer_i(:,:)
-    REAL(dp), INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
+    INTEGER, INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
     ! be replaced with this value
     ! if not provided all valid
     ! points will be packed to the
@@ -1946,7 +1946,7 @@ CONTAINS
     &                                  collector_buffer_i)
     ! dimension (:, global length); only required on root
     INTEGER, INTENT(INOUT) :: recv_buffer_i(:,:)
-    REAL(dp), INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
+    INTEGER, INTENT(IN), OPTIONAL :: fill_value ! if provided missing values will
     ! be replaced with this value
     ! if not provided all valid
     ! points will be packed to the
